나눗셈 정밀도 추가 & 히스토리 분수선 조정

+++ app/src/main/java/com/tistory/qlyh8/calculator/MainActivity.java
@@ -347,7 +347,7 @@ public class MainActivity extends AppCompatActivity {
         }
     }
 
-    public void setHistoryData(ArrayList<String> calc, String result, long[] fractionResult) {
+    public void setHistoryData(ArrayList<String> calc, String result, BigInteger[] fractionResult) {
         HistoryObject tempData;
         ArrayList<String> historyCalcList = new ArrayList<>();
         historyCalcList.addAll(calc);
